Remove "wildcards" from Firestore document event data
Internal discussions suggest this shouldn't be part of the event
data. (It isn't part of the payload in GCF events; it's provided
as a separate field in the root of the event body.) Client libraries
can provide manually-written code to parse document names instead.

Philosophically, this isn't part of the underlying event - it
depends on the event subscription (which contains the resource
template being used).