Wrap Cloud Audit Log event data in a log entry, to match reality on Cloud Run
It's not entirely obvious which fields *should* be included in LogEntry. They broadly fall into three categories:

Seen in examples, so included:

- log_name
- resource
- proto_payload
- insert_id
- timestamp
- receive_timestamp
- severity

Not seen in examples, but included as likely to be feasible:

- trace
- span_id

Not seen in examples, and not included as I don't expect to see them:

- text_payload
- json_payload
- http_request
- labels
- trace_sampled
- source_location

Before merging this, we should try to find out more information as to whether my guesses are correct.

Compatibility note: I've made proto_payload an AuditLog value, which is compatible with the JSON serialized form of a LogEntry, but would *not* be compatible with the binary serialized form, which would have an Any value instead.
If the eventing infrastructure wants to support protobuf binary serialization in the future, we'd probably need to introduce a new CloudEvent type for Cloud Audit Logs for that purpose.

The alternative is to make this an Any (in the same way it is in the original proto) - which makes it significantly harder to access the AuditLog data which is the genuinely useful part.